+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # MapTiler SDK Changelog
 
+## NEXT
+
+### ✨ Features and improvements
+- Adds the new `MapTilerAnimation` module and associated helpers for creating and managing animations, lerping between values and 'smoothing' arrays.
+- Adds the new `AnimatedRouteLayer` module and associated helpers for animating camera movement along GeoJSON features paths.
+
+### 🐛 Bug Fixes
+- None
+### ⚙️ Others
+- None
+
 ## 3.10.2
 ### ✨ Features and improvements
 - None